requirements.txt
aiogram>=3.15.0,<4.0.0
aiosqlite>=0.20.0,<0.21.0
httpx>=0.28.0,<0.29.0
beautifulsoup4>=4.12.0,<5.0.0
matplotlib>=3.9.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
playwright>=1.52.0,<2.0.0

